Modify ownCloudGui::slotOpenMainDialog to show the appropriate window(s)
The previous implementation only considered showing the Tray window, which may lead
to undesired behaviour:

- If there are no accounts configured, the wizard should be shown instead
- Other open windows would not be shown on second app invocation (Settings, ShareDialog)

This commit simply changes the call to ownCloudGui::slotTrayClicked, so if the user
starts the app a second time (e.g. clicking on the Dock icon), the outcome will be like
a click on the Tray icon.

Signed-off-by: Michael Schuster <michael@schuster.ms>